Check for data type equivalence one level higher, in ColumnComparator.
When doing a diff now, the two columns are not marked as having a different datatype, despite one being int8 and one being bigserial. This is intentional because both are autoincrement.
